Added more queries to Room database. Added more unit tests to Room database. Inject coroutine dispatchers. Use record id as Long type, not Int.

+++ app/src/main/java/com/dimowner/audiorecorder/v2/data/PrefsV2.kt
@@ -29,7 +29,7 @@ interface PrefsV2 {
 
     var askToRenameAfterRecordingStopped: Boolean
 
-    var activeRecordId: Int
+    var activeRecordId: Long
 
     val recordCounter: Long
     fun incrementRecordCounter()
